backend/requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
openai-whisper